mercedes-connected-car-adapter/pyproject.toml
[tool.poetry]
name = "mercedes-connected-car-adapter"
version = "0.1.0"
description = ""
authors = ["sionabraham <sion.abraham@tacto.ai>"]

[tool.poetry.dependencies]
python = "^3.9"
requests-oauthlib = "^1.3.0"
pytest-asyncio = "^0.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
